toolrental/farmer_django/Krishi-Yantra/backend/requirements.txt
# Backend stack for Krishi-Yantra (matches backend/settings.py)
Django>=5.2,<5.3
djangorestframework>=3.15
django-cors-headers>=4.4
djangorestframework-simplejwt>=5.3
Pillow>=10.0
